clients crud

+++ controllers/accountController.js
@@ -4,8 +4,9 @@ const regexFecha = /^\d{2}\/\d{2}\/\d{4}$/;
 exports.addAccount = async (req, res) => {
     try {
         const {expiration} = req.body;
+        const {userId} = req.query;
         if (!regexFecha.test(expiration)) throw Error("The expiration field must be a date");
-        const newAccount = await Account.create(req.body);
+        const newAccount = await Account.create({...req.body, userId});
         res.status(200).json(newAccount);
     } catch (err) {
         res.status(400).json({message: err.message});
@@ -14,7 +15,8 @@ exports.addAccount = async (req, res) => {
 
 exports.getAccounts = async (req, res) => {
     try {
-        const accounts = await Account.findAll();
+        const {userId} = req.query;
+        const accounts = await Account.findAll({where: {userId}});
         res.status(200).json(accounts);
     } catch (err) {
         res.status(400).json({message: err.message});
@@ -23,8 +25,9 @@ exports.getAccounts = async (req, res) => {
 
 exports.getAccountById = async (req, res) => {
     try {
+        const {userId} = req.query;
         const {id} = req.params;
-        const isValid = await Account.findByPk(id);
+        const isValid = await Account.findAll({where: {userId, id}});
         if(!isValid) throw Error("AccountId doesn't exist");
         const account = await Account.findByPk(id);
         res.status(200).json(account);
@@ -35,10 +38,11 @@ exports.getAccountById = async (req, res) => {
 
 exports.updateAccount = async (req, res) => {
     try {
+        const {userId} = req.query;
         const {id} = req.params;
-        const isValid = await Account.findByPk(id);
+        const isValid = await Account.findAll({where: {userId, id}});
         if(!isValid) throw Error("AccountId doesn't exist");
-        const account = await Account.update(req.body, { where: { id }});
+        const account = await Account.update(req.body, {where: {userId, id}});
         res.status(200).json({message: `${account} fields updated successfully`});
     } catch (err) {
         res.status(400).json({message: err.message});
@@ -47,10 +51,11 @@ exports.updateAccount = async (req, res) => {
 
 exports.deleteAccount = async (req, res) => {
     try {
+        const {userId} = req.query;
         const {id} = req.params;
-        const isValid = await Account.findByPk(id);
+        const isValid = await Account.findAll({where: {userId, id}});
         if(!isValid) throw Error("AccountId doesn't exist");
-        const account = await Account.destroy({ where: { id }});
+        const account = await Account.destroy({where: {userId, id}});
         res.status(200).json({message: "Account deleted successfully"});
     } catch (err) {
         res.status(400).json({message: err.message});
